server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
boto3==1.34.0
pydantic==2.11.3
python-multipart==0.0.22
jinja2==3.1.6
aiofiles==23.2.1
PyJWT==2.9.0
cryptography==46.0.5
requests==2.32.4
python-dotenv==1.0.0
# EAGLE additions
strands-agents[otel]>=1.0.0
anthropic>=0.40.0
# claude-agent-sdk>=0.1.19 — REMOVED: replaced by strands-agents (Phase 5 migration)
mcp>=1.5.0
python-docx>=1.1.0
openpyxl>=3.1.0
reportlab>=4.1.0
httpx==0.27.0
websockets>=12.0
# AgentCore SDK — Memory, Browser, Code Interpreter, Gateway, Identity, Policy, Runtime
bedrock-agentcore>=1.4.0
# OTEL observability (Phase 4 — trace spans for agent operations)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
# Langfuse trace visualization (OTEL-based, free tier 50K observations/month)
langfuse>=3.0.0
# Eval video recording: install separately with `pip install playwright>=1.40.0`
# Not included here — prevents unnecessary 46MB download in container builds
